pyproject.toml
[project]
name = "summony"
version = "0.1.0"
description = "Interfaces for simultaneously chatting with multiple LLMs."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "dataclasses-json>=0.6.7",
    "ipywidgets>=8.1.5",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.36.0",
]
ollama = [
    "ollama>=0.3.3",
]
openai = [
    "openai>=1.51.2",
]
google = [
    "google-generativeai>=0.8.3",
]
groq = [
    "groq>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
